Update README.md
I constructively think the README.md directions for cross-compiling for Windows with mxe is outdated and can be updated by adding cmake to the mxe tools (shown next to where gcc is) and invoking it directly. Note, as per the mxe master branch on 2020-06-09, direct use of the toolchain file is deprecated. Further note, the master branch is the recommended branch since stable was deprecated back in 2015, viz https://mxe.cc/#history.

Additionally, I constructively feel the README.md can be enhanced with a link to an example on how to produce a library for consumption by MSVC projects. I have added a link demonstrating how this can be accomplished by direct analogy to the proposed changes just before ### Testing

In summary, I propose appending a package name to one line, replacing two lines with an updated cmake command, and adding a line with an href at the end all within ### Building for Windows

Thank you for considering my proposed changes.